hack/services/user_manager_service/requirements.txt
django
djangorestframework
drf-yasg
